+++ src/main/java/insane96mcp/mobspropertiesrandomness/setup/Strings.java
@@ -8,5 +8,6 @@ public class Strings {
 		public static final String ON_ATTACK = MobsPropertiesRandomness.RESOURCE_PREFIX + "on_attack";
 		public static final String ON_ATTACKED = MobsPropertiesRandomness.RESOURCE_PREFIX + "on_attacked";
 		public static final String ON_DEATH = MobsPropertiesRandomness.RESOURCE_PREFIX + "on_death";
+		public static final String ON_TICK = MobsPropertiesRandomness.RESOURCE_PREFIX + "on_tick";
 	}
 }